Refactor docker build strategy
dumb-init will now be the default.  If somebody needs tini they can use
the tini flavor.

Three versions of nginx will now be available:  1.14, 1.16, 1.17.  This
ensures that an image with an acceptable nginx version is available, as
well as a base image using UBI 8, Fedora 31, and the official nginx base
image